Move monoid and semigroup invariant instances to invariant
