+++ core/src/main/scala/cats/instances/package.scala
@@ -21,14 +21,12 @@ package object instances {
   object list            extends ListInstances
   object long            extends LongInstances
   object map             extends MapInstances
-  object monoid          extends MonoidInstances
   object option          extends OptionInstances
   object order           extends OrderInstances
   object ordering        extends OrderingInstances
   object partialOrder    extends PartialOrderInstances
   object partialOrdering extends PartialOrderingInstances
   object queue           extends QueueInstances
-  object semigroup       extends SemigroupInstances
   object set             extends SetInstances
   object short           extends ShortInstances
   object stream          extends StreamInstances
